requirements.txt
aiodns==3.1.0
aiosignal==1.3.1
aiohttp==3.9.0b0
aiomysql==0.2.0
annotated-types==0.5.0
anyio==3.7.1
apscheduler==3.10.
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
discord-ext-menus==1.1
discord.py==2.3.2
discord-py-slash-command==3.0.3
discord-typings==0.7.0
DiscordUtils==1.3.4
emoji==2.8.0
fastapi==0.103.2
Flask==2.3.3
Flask-OAuthlib==0.9.6
frozenlist==1.4.0
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hmtai==3.2.31
hpack==4.0.0
hstspreload==2023.1.1
httpcore==0.18.0
httpx>=0.23.0
hyperframe==6.0.1
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
marblex.py==1.4.0
MarkupSafe==2.1.3
merakicommons==1.0.10
multidict==6.0.4
networkx==3.1
oauthlib==2.1.0
openai==0.28.0
packaging==23.1
Pillow==10.0.1
pip==23.3
pluggy==1.3.0
pycares==4.4.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyNaCl==1.5.0
pytest==7.4.2
pytz==2023.3.post1
raven==6.10.0
requests==2.31.0
requests-oauthlib==1.1.0
rfc3986==1.5.0
riotwatcher==3.2.5
schedule==1.2.0
sentinels==1.0.0
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
trivia.py==1.0.8
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
urllib3==2.0.7
utils==1.0.1
uvicorn==0.23.2
wcwidth==0.2.6
Werkzeug==3.0.1
yarl==1.9.2                                                                                                                                                                   